implemented a system for storing user settings via shared preferences + dynamic ui changes and custom views

+++ app/src/main/java/bg/zahov/app/data/Settings.kt
@@ -8,7 +8,7 @@ data class Settings(
     val distance: Units = Units.Normal,
     val soundEffects: Boolean = true,
     val theme: Theme = Theme.Dark,
-    val restTimer: Long = 30,
+    val restTimer: Int = 30,
     val vibration: Boolean = true,
     val soundSettings: Sound = Sound.SOUND_1,
     val updateTemplate: Boolean = true,
